+++ src/Gate.IO.Api/Spot/Stream/GateSpotStreamTrade.cs
@@ -41,7 +41,7 @@ public record GateSpotStreamTrade
     /// <summary>
     /// Order side
     /// </summary>
-    [JsonProperty("side")]
+    [JsonProperty("side"), JsonConverter(typeof(MapConverter))]
     public GateSpotOrderSide Side { get; set; }
 
     /// <summary>